More caret view consolidation, and add a todo list before we can merge
